batch class deserves its own file
