+++ lib/orel/table.rb
@@ -66,28 +66,6 @@ module Orel
       BatchQuery.new(select, batch, @heading, manager, table).results
     end
 
-    class Batch
-      def initialize(manager, heading, connection)
-        @connection = connection
-        @manager = manager
-        @heading = heading
-      end
-
-      def read_all
-        read
-      end
-
-      def read_batch(start, count)
-        @manager.take count
-        @manager.skip start
-        read
-      end
-
-      def read
-        @connection.execute(@manager.to_sql, @description || "#{self.class} Query #{@heading.name}").each(:as => :hash, :symbolize_keys => true)
-      end
-    end
-
     class Select
       attr_reader :batch_size
       attr_reader :batch_group
